Corrigir a ortografia da classe de erro
