+++ src/main/java/com/br/matchbook/models/Error.java
@@ -1,0 +1,37 @@
+package com.br.matchbook.models;
+
+public class Error {
+	private String message;
+	private String field;
+
+	public Error(String message, String field) {
+		this.field = field;
+		this.message = message;
+	}
+
+	public String getField() {
+		return field;
+	}
+
+	public void setField(String field) {
+		this.field = field;
+	}
+
+	public String getMessage() {
+		return message;
+	}
+
+	public void setMessage(String message) {
+		this.message = message;
+	}
+
+	@Override
+	public String toString() {
+		StringBuilder modelo = new StringBuilder();
+
+		modelo.append(this.field);
+		modelo.append(this.message);
+		return modelo.toString();
+		
+	}
+}
